Split CommonStorageBackend
changelog_begin
changelog_end
