+++ ledger/participant-integration-api/src/main/scala/platform/store/backend/oracle/OracleStorageBackend.scala
@@ -8,12 +8,17 @@ import anorm.SQL
 import com.daml.lf.data.Ref
 import com.daml.platform.store.backend.common.{
   AppendOnlySchema,
-  CommonStorageBackend,
   CompletionStorageBackendTemplate,
+  ConfigurationStorageBackendTemplate,
   ContractStorageBackendTemplate,
+  DataSourceStorageBackendTemplate,
+  DeduplicationStorageBackendTemplate,
   EventStorageBackendTemplate,
   EventStrategy,
+  IngestionStorageBackendTemplate,
   InitHookDataSourceProxy,
+  PackageStorageBackendTemplate,
+  ParameterStorageBackendTemplate,
   PartyStorageBackendTemplate,
   QueryStrategy,
   Timestamp,
@@ -36,7 +41,12 @@ import javax.sql.DataSource
 
 private[backend] object OracleStorageBackend
     extends StorageBackend[AppendOnlySchema.Batch]
-    with CommonStorageBackend[AppendOnlySchema.Batch]
+    with DataSourceStorageBackendTemplate
+    with IngestionStorageBackendTemplate[AppendOnlySchema.Batch]
+    with ParameterStorageBackendTemplate
+    with ConfigurationStorageBackendTemplate
+    with PackageStorageBackendTemplate
+    with DeduplicationStorageBackendTemplate
     with EventStorageBackendTemplate
     with ContractStorageBackendTemplate
     with CompletionStorageBackendTemplate
